layout/inspector/tests/test_bug536379.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=536379
-->
<head>
  <title>Test for Bug 536379</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <link rel="stylesheet" type="text/css" href="data:text/css,p { color: green }">
  <link rel="stylesheet" type="text/css" href="data:text/css,p { color: green }">
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=536379">Mozilla Bug 536379</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 536379 **/

netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

const CI = Components.interfaces;
const CC = Components.classes;

var domUtils =
  CC["@mozilla.org/inspector/dom-utils;1"].getService(CI.inIDOMUtils);
var rules = domUtils.getCSSStyleRules(document.getElementById("display"));
var firstPRule =
  rules.GetElementAt(rules.Count() - 2).QueryInterface(CI.nsIDOMCSSStyleRule);
firstPRule.style.removeProperty("color");
ok(true, "should not crash");

var links = document.getElementsByTagName("link");
is(links.length, 3, "links.length");
is(firstPRule.parentStyleSheet, links[1].sheet, "sheet match for first P rule");
var secondPRule =
  rules.GetElementAt(rules.Count() - 1).QueryInterface(CI.nsIDOMCSSStyleRule);
is(secondPRule.parentStyleSheet, links[2].sheet, "sheet match for second P rule");
is(links[1].href, links[2].href, "links should have same href");
isnot(links[1].sheet, links[2].sheet, "links should have different sheets");
isnot(firstPRule, secondPRule, "rules should be different");
isnot(firstPRule.cssText, secondPRule.cssText, "text should be different since property was removed from one");

</script>
</pre>
</body>
</html>
